\name{aa.prop.list}
\alias{aa.prop.list}
\docType{data}
\title{
Amino Acid Properties
}
\description{
Amino Acid Properties

}
\format{
  A data frame with 20 observations on the following 13 variables.
  \describe{
    \item{\code{Symbol}}{a chacracter vector with 20 values: A through Y}
    \item{\code{AA_Name}}{a chacracter vector with 20 values: Alanine through Tyrosine}
    \item{\code{AA_Symbol}}{a chacracter vector with 20 values: Ala through Tyr}
    \item{\code{Surface_Area_Chothia}}{a numeric vector}
    \item{\code{Residue_Volume_Zamayatin}}{a numeric vector}
    \item{\code{Bulkiness_Jones}}{a numeric vector}
    \item{\code{Polarity_Jones}}{a numeric vector}
    \item{\code{Refractivity_Jones}}{a numeric vector}
    \item{\code{Hydrophobicity_Engleman}}{a numeric vector}
    \item{\code{Hydrophobicity_Prabha}}{a numeric vector}
    \item{\code{Hydrophilicity_Hopp}}{a numeric vector}
    \item{\code{Hydrophilicity_Levitt}}{a numeric vector}
    \item{\code{RelMutability_Jones}}{a numeric vector}
  }
}
