\name{chi.norm}
\alias{chi.norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A Transformation of Chi-squared Random Variable
}
\description{
A transformation of Chi-squared random variable to make it normal like.
}
\usage{
chi.norm(q, v)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{q}{
numeric. A random variable following chi-squared distribution
}
  \item{v}{
numeric. A random variable following normal distribution
}
}
