\name{cloud9}
\alias{cloud9}
\docType{data}
\title{
9-Component Mixture Dirichlet Prior for Protein Sequences
}
\description{
9-Component Mixture Dirichlet Prior for Protein Sequences
}
\format{
List of 2. The alpha element is a 9 by 20 matrix, where each row represents one Dirichlet distribution of 20 dimensions. 
The mix.coef element contains the mixing probability, a vector 0f 9 numbers that add up to 1.
}
