\name{sim.liu.2008}
\alias{sim.liu.2008}
\alias{sim.liu.2007}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulate sDataset
}
\description{
Per Liu et al (2008) and Liu et al (2007)
}
\usage{
sim.liu.2008(n, a, seed = NULL)
sim.liu.2007(n, a, seed = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
sample size
}
  \item{a}{
numeric. If a is 0, then the data is used to study size, otherwise power
}
  \item{seed}{
optional random number generator seed
}
}
